// frontend/src/api/models/listingSourceStats.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Bostad API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingSources } from './listingSources';
import type { ListingSourceStatsLoggedIn } from './listingSourceStatsLoggedIn';

/**
 * Per-source scrape statistics exposed to the frontend.
 */
export interface ListingSourceStats {
  source: ListingSources;
  name: string;
  globalUrl: string;
  loggedIn?: ListingSourceStatsLoggedIn;
  numListings?: number;
  numErrors?: number;
}